+++ src/server/game/Entities/Player/Player.h
@@ -2232,8 +2232,8 @@ class TC_GAME_API Player : public Unit, public GridObject<Player>
         void ClearInventory();
         void ApplyVirtualItemLegendayEffects(Item* item);
 
-        float GetMagicFind() { return magicFind; }
-        void SetMagicFind(float amount) { magicFind = amount; }
+        uint32 GetMagicFind() { return magicFind; }
+        void SetMagicFind(uint32 amount) { magicFind = amount; }
 
         WorldLocation GetPortalLocation() { return _portalLocation; }
         void SetPortalLocation(WorldLocation loc) { _portalLocation = loc; }
@@ -2583,7 +2583,7 @@ class TC_GAME_API Player : public Unit, public GridObject<Player>
 
         uint32 talent_level;
         bool m_canTeleport;
-        float magicFind;
+        uint32 magicFind;
 
         WorldLocation _portalLocation;
         uint8 subClass;
