+++ src/server/game/Globals/VirtualItemMgr.h
@@ -55,7 +55,7 @@ struct VirtualItemTemplate : ItemTemplate
         legendarySeed = 0;
         legendaryId = 0;
         customFlags = 0;
-        generatedMagicFind = 0.f;
+        generatedMagicFind = 0;
         honePct = 0.f;
         //Bonding = BIND_WHEN_PICKED_UP; // All items MUST be bound on pickup so they can not be mailed and thus failing some cleanup and memory management
     }
@@ -92,7 +92,7 @@ struct VirtualItemTemplate : ItemTemplate
     // misc
     uint32 legendaryId;
     uint32 customFlags;
-    float generatedMagicFind;
+    uint32 generatedMagicFind;
     float honePct;
     inline bool HasFlag(VirtualItemFlags flag) const { return (customFlags & flag) != 0; }
 };
@@ -121,7 +121,7 @@ struct VirtualModifier
         vLvlMod = 0;
         displayId = 0;
         nameOverride = "";
-        magicFind = 0.f;
+        magicFind = 0;
         statPoolPctModifier = 0.f;
         subclass = 0;
         generateSet = false;
@@ -150,7 +150,7 @@ struct VirtualModifier
     uint32 vLvlMod;
     uint32 displayId;
     std::string nameOverride;
-    float magicFind;
+    uint32 magicFind;
     float statPoolPctModifier;
     uint8 subclass;
     bool generateSet;
